Moved test back to corresponding domains.
