+++ pkg/starkapi/assets_api_test.go
